+++ src/queries/auth.ts
@@ -1,0 +1,27 @@
+import { paths } from '../routes/paths'
+import { useMutation } from '@tanstack/react-query'
+import { login, signUp } from 'api'
+import { queryClient } from 'queries'
+import { useNavigate } from 'react-router-dom'
+import { tokenRepository } from 'utils'
+
+export const useLoginMutation = () => {
+  const navigate = useNavigate()
+  return useMutation(login, {
+    onSuccess: (value) => {
+      tokenRepository.value = value.token
+      queryClient.invalidateQueries(['todos'])
+      navigate(paths.root)
+    },
+  })
+}
+export const useSignUpMutation = () => {
+  const navigate = useNavigate()
+  return useMutation(signUp, {
+    onSuccess: (value) => {
+      tokenRepository.value = value.token
+      queryClient.invalidateQueries(['todos'])
+      navigate(paths.root)
+    },
+  })
+}
